sam_droplet/requirements_server.txt
flask>=2.3.0
flask-cors>=4.0.0
waitress>=2.1.0
gunicorn>=21.0.0
pillow>=9.5.0
numpy>=1.24.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
segment-anything>=1.0
timm>=0.9.0
scipy>=1.10.0
scikit-image>=0.20.0
matplotlib>=3.7.0
scikit-learn>=1.3.0 